WIP refactor directory
