+++ lib/array.rb
