--- models/integer.rb
@@ -1,5 +1,0 @@
-class Integer
-  def args
-    [self]
-  end
-end
