--- models/string.rb
@@ -1,5 +1,0 @@
-class String
-  def args
-    [self]
-  end
-end
